added routing and controllers
