+++ controllers/users.js
@@ -1,0 +1,105 @@
+const express = require('express');
+const jwt = require('jsonwebtoken');
+const passport = require('passport');
+const User = require('../models/user');
+require('dotenv').config();
+
+const secret = process.env.SECRET;
+
+const auth = (req, res, next) => {
+  passport.authenticate(
+    'jwt',
+    { session: false },
+    (err, user) => {
+      if (!user || err) {
+        return res.status(401).json({
+          status: 'error',
+          code: 401,
+          message: 'Unauthorized',
+          data: 'Unauthorized',
+        });
+      }
+      req.user = user;
+      next();
+    }
+  )(req, res, next);
+};
+
+const signup = async (req, res, next) => {
+  const { username, email, password } = req.body;
+  const user = await User.findOne({ email });
+  if (user) {
+    return res.status(409).json({
+      status: 'error',
+      code: 409,
+      message: 'Email is already in use',
+      data: 'Conflict',
+    });
+  }
+  try {
+    const newUser = new User({ username, email });
+    newUser.setPassword(password);
+    await newUser.save();
+    res.status(201).json({
+      status: 'success',
+      code: 201,
+      data: {
+        message: 'Registration successful',
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const login = async (req, res, next) => {
+  const { email, password } = req.body;
+  const user = await User.findOne({ email });
+
+  if (!user || !user.validPassword(password)) {
+    return res.status(400).json({
+      status: 'error',
+      code: 400,
+      message: 'Incorrect login or password',
+      data: 'Bad request',
+    });
+  }
+
+  const payload = {
+    id: user.id,
+    username: user.username,
+  };
+
+  const token = jwt.sign(payload, secret, {
+    expiresIn: '1h',
+  });
+  res.json({
+    status: 'success',
+    code: 200,
+    data: {
+      token,
+    },
+  });
+};
+
+const logout = (req, res, next) => {};
+
+const current =
+  (auth,
+  (req, res, next) => {
+    const { username } = req.user;
+    res.json({
+      status: 'success',
+      code: 200,
+      data: {
+        message: `Authorization was successful: ${username}`,
+      },
+    });
+  });
+
+module.exports = {
+  signup,
+  login,
+  logout,
+  current,
+};
