+++ routes/api/users.js
@@ -1,100 +1,10 @@
 const express = require('express');
 const router = express.Router();
-const jwt = require('jsonwebtoken');
-const passport = require('passport');
-const User = require('../../models/user');
-require('dotenv').config();
+const authController = require('../../controllers/users');
 
-const secret = process.env.SECRET;
-
-const auth = (req, res, next) => {
-  passport.authenticate(
-    'jwt',
-    { session: false },
-    (err, user) => {
-      if (!user || err) {
-        return res.status(401).json({
-          status: 'error',
-          code: 401,
-          message: 'Unauthorized',
-          data: 'Unauthorized',
-        });
-      }
-      req.user = user;
-      next();
-    }
-  )(req, res, next);
-};
-
-router.post('/users/login', async (req, res, next) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-
-  if (!user || !user.validPassword(password)) {
-    return res.status(400).json({
-      status: 'error',
-      code: 400,
-      message: 'Incorrect login or password',
-      data: 'Bad request',
-    });
-  }
-
-  const payload = {
-    id: user.id,
-    username: user.username,
-  };
-
-  const token = jwt.sign(payload, secret, {
-    expiresIn: '1h',
-  });
-  res.json({
-    status: 'success',
-    code: 200,
-    data: {
-      token,
-    },
-  });
-});
-
-router.post('/users/signup', async (req, res, next) => {
-  const { username, email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (user) {
-    return res.status(409).json({
-      status: 'error',
-      code: 409,
-      message: 'Email is already in use',
-      data: 'Conflict',
-    });
-  }
-  try {
-    const newUser = new User({ username, email });
-    newUser.setPassword(password);
-    await newUser.save();
-    res.status(201).json({
-      status: 'success',
-      code: 201,
-      data: {
-        message: 'Registration successful',
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-//TODO:
-router.post('/users/logout', (req, res, next) => {});
-
-router.get('/users/current', auth, (req, res, next) => {
-  const { username } = req.user;
-  res.json({
-    status: 'success',
-    code: 200,
-    data: {
-      message: `Authorization was successful: ${username}`,
-    },
-  });
-});
+router.post('/users/login', authController.login);
+router.post('/users/logout', authController.logout);
+router.post('/users/register', authController.signup);
+router.post('/users/current', authController.current);
 
 module.exports = router;
